requirements.txt
# Polymarket Trading Bot Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# Core Dependencies
# =============================================================================

# Ethereum libraries for wallet operations
web3>=6.0.0                    # Polygon RPC interactions
eth-account>=0.5.0             # Wallet and signing operations
eth-abi>=4.0.0                 # ABI encoding for contracts

# Cryptography for private key encryption
cryptography>=41.0.0           # PBKDF2 + Fernet encryption

# Configuration
pyyaml>=6.0                    # YAML config file support
python-dotenv>=1.0.0           # .env file loading

# HTTP requests
requests>=2.28.0               # API calls

# WebSocket for real-time data
websockets>=12.0               # WebSocket client for market data

# =============================================================================
# Polymarket API Clients (Optional - for advanced usage)
# =============================================================================

# Official Polymarket clients
py-builder-relayer-client>=0.1.0  # For gasless transactions (Safe wallets)
py-builder-signing-sdk>=0.1.0    # For Builder API signing

# =============================================================================
# Development & Testing
# =============================================================================

pytest>=7.0.0                  # Test framework
pytest-asyncio>=0.21.0         # Async test support
pytest-cov>=4.0.0              # Code coverage
